consul_service/kubernetes_discovery/requirements.txt
flask==2.3.3
requests==2.31.0
flask-cors==4.0.0
pyjwt==2.8.0
bcrypt==4.0.1
uuid==1.30
python-dotenv==1.0.0
etcd3==0.12.0